Random action functionality acquired
The basic setup to test the functionality of the
agent when it randomly choses an action rather than
from learning an action. This is done to observe the
initial results before any reinforced learning is
applied.

Also added a function in the LearningAgent class
called update_state where i've started to create the
function that will take the agents state and apply the
learning technique using a Q-value table.