Completed the update_state function and started 3 others
Completed the update_state function which is used to
figure out what state it's in based on the input
variables and then return that state as a list.

Also started 3 functions:
- 1: lookup_actions() used to find the q-values for each of 4 possible
     actions and then decide based upon whats given a higher q-value
     which action to take.
- 2: update_policy() used to update the q-value table based upon a
     state, action, and reward.
- 3: q_values() used to store the q-value table that lookup_actions()
     needs to lookup q-values and that update_policy() needs to
     update the q-values.

Also added some more explanatory text at beginning as well as
got rid of some less than useful information.