Changed static variables to state variables and initialized q table during initialization
Took the static variables in the Agent class
and turned them into state variables in case of
the addition of more agents.
Also switched the initialization of the q table
from thee update() function to the initialization
function.
Also switched the variable cross_traffic to be two
different variables, left and right, to help it learn
to go right on red if theres no traffic on the left side
while at the same time allowing it to use London traffic rules
if it happened to switch to that.